requirements.txt
# Core Data Science Libraries
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Machine Learning
xgboost>=1.7.0
lightgbm>=4.0.0
imbalanced-learn>=0.11.0

# Model Persistence
joblib>=1.3.0

# API and Web Framework
flask>=2.3.0
flask-cors>=4.0.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Data Validation
great-expectations>=0.17.0
